Move the LaTeX validation from models.py to forms.py
